Refactor base transaction type component
