+++ front-end/src/app/shared/components/transaction-type-base/form.behaviors.ts
@@ -1,0 +1,186 @@
+import { Component, Input, OnDestroy, OnInit } from '@angular/core';
+import { AbstractControl, FormBuilder, FormGroup } from '@angular/forms';
+import { Router } from '@angular/router';
+import { TransactionType } from 'app/shared/models/transaction-types/transaction-type.model';
+import { MemoText } from 'app/shared/models/memo-text.model';
+import { SchATransaction, ScheduleATransactionTypes } from 'app/shared/models/scha-transaction.model';
+import {
+  NavigationAction,
+  NavigationControl,
+  NavigationDestination,
+  TransactionNavigationControls,
+} from 'app/shared/models/transaction-navigation-controls.model';
+import { Transaction } from 'app/shared/models/transaction.model';
+import { FecDatePipe } from 'app/shared/pipes/fec-date.pipe';
+import { ContactService } from 'app/shared/services/contact.service';
+import { TransactionService } from 'app/shared/services/transaction.service';
+import { ValidateService } from 'app/shared/services/validate.service';
+import { LabelUtils, PrimeOptions } from 'app/shared/utils/label.utils';
+import { ConfirmationService, MessageService, SelectItem } from 'primeng/api';
+import { BehaviorSubject, combineLatestWith, Observable, of, startWith, Subject, switchMap, takeUntil } from 'rxjs';
+import { Contact, ContactFields, ContactTypeLabels, ContactTypes } from '../../models/contact.model';
+import { MemoBehaviors } from './memo.behaviors';
+
+export class FormBehaviors {
+  static onInit(
+    form: FormGroup,
+    validateService: ValidateService,
+    transactionType: TransactionType | undefined,
+    contactId$: Subject<string>,
+    contributionPurposeDescriptionLabel: string
+  ): void {
+    // Initialize validation tracking of current JSON schema and form data
+    validateService.formValidatorSchema = transactionType?.schema;
+    validateService.formValidatorForm = form;
+
+    // Override contact type options if present in transactionType
+    if (transactionType && transactionType.contactTypeOptions) {
+      this.contactTypeOptions = LabelUtils.getPrimeOptions(ContactTypeLabels, transactionType.contactTypeOptions);
+    }
+
+    // Intialize form values
+    function isExisting(transaction: Transaction | undefined) {
+      return !!transaction?.id;
+    }
+
+    if (isExisting(transactionType?.transaction)) {
+      const txn = { ...transactionType?.transaction } as SchATransaction;
+      form.patchValue({ ...txn });
+
+      MemoBehaviors.patchMemoText(transactionType, form);
+
+      form.get('entity_type')?.disable();
+      contactId$.next(txn.contact_id || '');
+    } else {
+      this.resetForm();
+      form.get('entity_type')?.enable();
+      this.contactId$.next('');
+    }
+
+    form
+      .get('entity_type')
+      ?.valueChanges.pipe(takeUntil(this.destroy$))
+      .subscribe((value: string) => {
+        if (value === ContactTypes.INDIVIDUAL || value === ContactTypes.CANDIDATE) {
+          form.get('contributor_organization_name')?.reset();
+        }
+        if (value === ContactTypes.ORGANIZATION || value === ContactTypes.COMMITTEE) {
+          form.get('contributor_last_name')?.reset();
+          form.get('contributor_first_name')?.reset();
+          form.get('contributor_middle_name')?.reset();
+          form.get('contributor_prefix')?.reset();
+          form.get('contributor_suffix')?.reset();
+          form.get('contributor_employer')?.reset();
+          form.get('contributor_occupation')?.reset();
+        }
+      });
+
+    form
+      ?.get('contribution_aggregate')
+      ?.valueChanges.pipe(takeUntil(this.destroy$))
+      .subscribe(() => {
+        form.get('contributor_employer')?.updateValueAndValidity();
+        form.get('contributor_occupation')?.updateValueAndValidity();
+      });
+
+    const previous_transaction$: Observable<Transaction | undefined> =
+      form.get('contribution_date')?.valueChanges.pipe(
+        startWith(form.get('contribution_date')?.value),
+        combineLatestWith(contactId$),
+        switchMap(([contribution_date, contactId]) => {
+          return this.transactionService.getPreviousTransaction(transactionType, contactId, contribution_date);
+        })
+      ) || of(undefined);
+    form
+      .get('contribution_amount')
+      ?.valueChanges.pipe(
+        startWith(form.get('contribution_amount')?.value),
+        combineLatestWith(previous_transaction$),
+        takeUntil(this.destroy$)
+      )
+      .subscribe(([contribution_amount, previous_transaction]) => {
+        const previousAggregate = +((previous_transaction as SchATransaction)?.contribution_aggregate || 0);
+        form.get('contribution_aggregate')?.setValue(+contribution_amount + previousAggregate);
+      });
+
+    const contribution_amount_schema = this.transactionType?.schema.properties['contribution_amount'];
+    if (contribution_amount_schema?.exclusiveMaximum === 0) {
+      this.negativeAmountValueOnly = true;
+      form
+        .get('contribution_amount')
+        ?.valueChanges.pipe(takeUntil(this.destroy$))
+        .subscribe((contribution_amount) => {
+          if (typeof contribution_amount === 'number' && contribution_amount > 0) {
+            form.patchValue({ contribution_amount: -1 * contribution_amount });
+          }
+        });
+    }
+
+    if (this.transactionType?.generatePurposeDescriptionLabel) {
+      this.contributionPurposeDescriptionLabel = this.transactionType.generatePurposeDescriptionLabel();
+    }
+  }
+
+  static getPayloadTransaction(
+    transactionType: TransactionType | undefined,
+    validateService: ValidateService,
+    form: FormGroup,
+    formProperties: string[]
+  ) {
+    let formValues = validateService.getFormValues(form, formProperties);
+    if (transactionType) formValues = MemoBehaviors.retrieveMemoText(transactionType, form, formValues);
+
+    const payload: SchATransaction = SchATransaction.fromJSON({
+      ...transactionType?.transaction,
+      ...formValues,
+    });
+    payload.contact_id = payload.contact?.id;
+
+    if (payload.children) {
+      payload.children = payload.updateChildren();
+    }
+
+    let fieldsToValidate: string[] = validateService.getSchemaProperties(transactionType?.schema);
+    // Remove properties that are populated in the back-end from list of properties to validate
+    fieldsToValidate = fieldsToValidate.filter(
+      (p) =>
+        ![
+          'transaction_id',
+          'donor_committee_name',
+          'back_reference_tran_id_number',
+          'back_reference_sched_name',
+        ].includes(p)
+    );
+    payload.fields_to_validate = fieldsToValidate;
+    function addFieldsToValidate(transaction: Transaction) {
+      transaction.fields_to_validate = fieldsToValidate;
+      if (transaction.children) {
+        transaction.children.forEach((transaction) => addFieldsToValidate(transaction));
+      }
+    }
+    addFieldsToValidate(payload);
+    return payload;
+  }
+
+  static resetForm(form: FormGroup, transactionType: TransactionType | undefined) {
+    this.formSubmitted = false;
+    form.reset();
+    form.markAsPristine();
+    form.markAsUntouched();
+
+    form.patchValue({
+      entity_type: this.contactTypeOptions[0]?.code,
+      contribution_aggregate: '0',
+      memo_code: this.getMemoCodeConstant(transactionType),
+      contribution_purpose_descrip: transactionType?.generatePurposeDescription?.() || '',
+    });
+  }
+
+  static getMemoCodeConstant(transactionType?: TransactionType): boolean | undefined {
+    /** Look at validation schema to determine if the memo_code has a constant value.
+     * If there is a constant value, return it, otherwise undefined
+     */
+    const memoCodeSchema = transactionType?.schema.properties['memo_code'];
+    return memoCodeSchema?.const as boolean | undefined;
+  }
+}
