+++ front-end/src/app/shared/components/transaction-type-base/transaction-type-x2-base.component.spec.ts
